cc: Only report calc_draw_props_time.
The perf dashboard is configured to monitor every trace output by the
cc_perftests. Here, we were outputting two traces that represent the exact
same thing. This only causes more work for the perf sheriffs. So I'm cutting it
down to one.

Incidentally, this eliminates a bug where the dashboard thought that a higher
count was a regression.

This also improves the precision by supporting decimals.

BUG=313537

Review URL: https://codereview.chromium.org/56573002

git-svn-id: http://src.chromium.org/chrome/trunk/src/testing@233775 4ff67af0-8c30-449e-8e8b-ad334ec8d88c