Get iossim to compile with Xcode 6.
Xcode 6 moved the location of the frameworks that iossim depends on and
added a few new dependecies so new header files need to be generated by
the gyp target.

Note that this CL gets iossim to compile but iossim will not launch the
Xcode 6 simulator.

BUG=381213
NOTRY=true
TBR=rohitrao@chromium.org

Review URL: https://codereview.chromium.org/336773002

git-svn-id: http://src.chromium.org/chrome/trunk/src/testing@278786 4ff67af0-8c30-449e-8e8b-ad334ec8d88c