In iossim, use default home directory if no -u override is given.
With Xcode 6, the home directory of the iOS Simulator and iossim
processes must match so that the iOS Simulator can find the SimDevice
object created by iossim. Since the existing behavior of iossim is to
create a new, random, home directory on each run (if no -u value is
specified on the command line), the home directories will be  mismatched
if iossim is launched multiple times without killing the iOS Simulator
in between runs.

The fix here is to only override the home directory in iossim if -u
is specified. Since this will often reuse the default SimDeviceSet
(at ~/Library/Developer/CoreSimulator/Devices), functionality
is also changed to reuse an existing SimDevice if possible (though
iossim will create a SimDevice if necessary).

BUG=None
TBR=rohitrao

Review URL: https://codereview.chromium.org/382563002

git-svn-id: http://src.chromium.org/chrome/trunk/src/testing@282139 4ff67af0-8c30-449e-8e8b-ad334ec8d88c